Add support for Fedora and Fedora CoreOS
Based on RHEL8 scripts. Unlike RHEL8, Fedora is highly-volatile and
packages expires so fast. It is possible that kernel installed on host
a week ago can be already pruned from the repositories. In this case
container will not be able to download kernel-headers, kernel-core and
kernel-devel packages. We workaround this problem by trying to download
old versions of packages directly from Koji (Fedora CI).

Tested on:

- Fedora 32 (kernel 5.6.19-300.fc32.x86_64)
- Fedora CoreOS 32.20200619.20.0 (kernel 5.6.19-300.fc32.x86_64)

Signed-off-by: Roman Tsisyk <rtsisyk@fedoraproject.org>